feat(ai): tags
