+++ apps/egghead/src/app/(content)/posts/_components/edit-post-form-metadata.tsx
@@ -5,7 +5,6 @@ import { useRouter } from 'next/navigation'
 import { PostPlayer } from '@/app/(content)/posts/_components/post-player'
 import { reprocessTranscript } from '@/app/(content)/posts/[slug]/edit/actions'
 import { NewLessonVideoForm } from '@/components/resources-crud/new-lesson-video-form'
-import AdvancedTagSelector from '@/components/resources-crud/tag-selector'
 import { env } from '@/env.mjs'
 import { useTranscript } from '@/hooks/use-transcript'
 import {
@@ -46,6 +45,7 @@ import {
 import { useSocket } from '@coursebuilder/ui/hooks/use-socket'
 import { MetadataFieldState } from '@coursebuilder/ui/resources-crud/metadata-fields/metadata-field-state'
 import { MetadataFieldVisibility } from '@coursebuilder/ui/resources-crud/metadata-fields/metadata-field-visibility'
+import AdvancedTagSelector from '@coursebuilder/ui/resources-crud/tag-selector'
 
 import { MetadataFieldAccess } from './metadata-field-access'
 import { PostUploader } from './post-uploader'
@@ -240,7 +240,7 @@ export const PostMetadataFormFields: React.FC<{
 				<AdvancedTagSelector
 					availableTags={tags}
 					selectedTags={post?.tags?.map((tag) => tag.tag) ?? []}
-					onTagSelect={async (tag: EggheadTag) => {
+					onTagSelect={async (tag: { id: string }) => {
 						await addTagToPost(post.id, tag.id)
 					}}
 					onTagRemove={async (tagId: string) => {
